Starting small restructure because of file length
